split amp/bento accordion tests
